Fix JavaFX tests hanging in headless mode.
Before, the initialization logic had a flaw. If a call to one API
failed, another API would be attempted. However, this approach is
problematic: if the first call failed with an exception for some
reason, it would leave JavaFX in a broken state where a flag would
imply that the system is being initialized. Subsequent calls would
then proceed to wait forever for the initialization to complete.

Setting a flag to `true` in a call to `startup` forces JavaFX to
fail with an exception when duplicate calls are detected. It seems
that in the older version this behavior was the default.